[WIP] views: move to nda for array views
NDA view data() calls inherit the const qualifier from the container, not the inner
type. mdspan inherited it from the inner type, not the qualifier on the container. This means
there is no easy way to use an rvalue reference (such as const &myview) to get editable data
from the view on the callee. Using the 'tree.proxy_view()' type calls on the caller side would
have to be bound to an lvalue first, removing their utility and make the code more verbose and
confusing. I could use forwarding references, which are a huge pain in the ass, if I was
willing to deal with these annoyances

1. They require the type to be deduced, which was generally not the case (it essentially requires
the argument to be 'T&&' with no qualifiers, and typically my template parameter is an inner
type, such as float/double, not ndview<T, dim>). This reduces the intent of the argument type
considerably, unless I make manual typedefs for each kind of view

2. DMK is not a header library, instead using separate compilation units for various
functions. Linking is even more of a pain in the butt because now I need to ensure the types are
perfect for all arguments. Since the compiler is deducing the type, and will just compile
whatever it sees in the template instantiation with no restriction, so long as it's valid code.
If I try to call that function from another compilation unit, I have to be 100% it will deduce
to the exact same function, otherwise I will get a linking error.

3. I'll have twice as many template instantiations, because I need to handle rvalues and lvalues
separately. I already have to do double/float separately, so now for every function I have to
write it out 4 times. Gross.

Solution: Views are small and non-owning. They don't need to be "moved", and never need to
worry about thread synchronization (excluding the underlying data). Just pass them by
value. Remove all inner type qualifiers. This is usually sufficient.

Issues still present: In some cases I need to construct a view on const data, but then pass it
to a function that is expecting a const view (which implies const data). But my functions all
expect views with a non-const inner type. NDA won't let you construct a non-const inner typed
container with a const pointer, even though the container itself is const -- so for now I
const-cast it away until I find a nicer way to deal with this.